+++ cognite/client/utils/__init__.py
@@ -6,8 +6,8 @@ import sys
 from cognite.client.utils._time import (
     MAX_TIMESTAMP_MS,
     MIN_TIMESTAMP_MS,
-    datetime_to_gql_timestamp,
     datetime_to_ms,
+    datetime_to_ms_iso_timestamp,
     ms_to_datetime,
     timestamp_to_ms,
 )
@@ -26,5 +26,5 @@ __all__ = [
     "datetime_to_ms",
     "ms_to_datetime",
     "timestamp_to_ms",
-    "datetime_to_gql_timestamp",
+    "datetime_to_ms_iso_timestamp",
 ]